feat!: Change methods to take HandlerFuncs instead of Handlers
I almost exclusively work with http.HandlerFuncs in practice and since
any http.Handler can be used by passing http.Handler.ServeHTTP, it's
easier for me to continue working in that fashion.

BREAKING CHANGE: This commit changes the signatures of the GET and POST
methods and is thus not backwards compatible, but we are not at a 1.x+
release yet, so this isn't an issue.